+++ src/Metrics/Data/MetricCollection.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace PipServices3.Azure.Metrics.Data
+{
+    public class MetricCollection
+    {
+        public IList<Metric> Value { get; set; } = new List<Metric>();
+    }
+}