+++ src/Metrics/Data/Unit.cs
@@ -1,0 +1,13 @@
+﻿namespace PipServices3.Azure.Metrics.Data
+{
+    public enum Unit
+    {
+        Count = 0,
+        Bytes = 1,
+        Seconds = 2,
+        CountPerSecond = 3,
+        BytesPerSecond = 4,
+        Percent = 5,
+        MilliSeconds = 6
+    }
+}