+++ src/Persistence/Data/ExcludedPathEntity.cs
@@ -1,0 +1,11 @@
+﻿using System.Runtime.Serialization;
+
+namespace PipServices3.Azure.Persistence.Data
+{
+    [DataContract]
+    public class ExcludedPathEntity
+    {
+        [DataMember(Name = "path")]
+        public string Path { get; set; }
+    }
+}