+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Azure specific components for Pip.Services in .NET Changelog
 
-## <a name="3.0.0-3.0.6"></a> 3.0.0-3.0.6 (2019-04-16)
+## <a name="3.0.0-3.0.7"></a> 3.0.0-3.0.7 (2019-10-02)
 
 ### Features
+* **lock** Add CloudStorageLock
+* **metrics** Add CosmosDbMetricsService
+* **persistence** Add CosmosMongoDbPartitionPersistence
 * **auth** Extended KeyVault client
 
 ### Breaking Changed